deleted boardpins.json and added boardPin function to pinData, attempting to get axios get to pull down correct info
