+++ db/!base.json
@@ -1,6 +1,5 @@
 {
   "users": "cool",
   "boards": "boards",
-  "pins": "pins",
-  "boardPins": "boardPins"
+  "pins": "pins"
 }
