--- db/boardPins.json
@@ -1,46 +1,0 @@
-{
-  "boardPins1": {
-    "pinId": "pin1",
-    "boardId": "board1" 
-  },
-  "boardPins2": {
-    "pinId": "pin2",
-    "boardId": "board1"
-  },
-  "boardPins3": {
-    "pinId": "pin3",
-    "boardId": "board1"
-  },
-  "boardPins4": {
-    "pinId": "pin4",
-    "boardId": "board1"
-  },
-  "boardPins5": {
-    "pinId": "pin4",
-    "boardId": "board2"
-  },
-  "boardPins6": {
-    "pinId": "pin5",
-    "boardId": "board2"
-  },
-  "boardPins7": {
-    "pinId": "pin6",
-    "boardId": "board2"
-  },
-  "boardPins8": {
-    "pinId": "pin9",
-    "boardId": "board2"
-  },
-  "boardPins9": {
-    "pinId": "pin9",
-    "boardId": "board3"
-  },
-  "boardPins10": {
-    "pinId": "pin7",
-    "boardId": "board3"
-  },
-  "boardPins11": {
-    "pinId": "pin8",
-    "boardId": "board3"
-  }
-}
