+++ src/javascripts/helpers/data/pinData.js
@@ -17,4 +17,14 @@ const getPins = () => new Promise((resolve, reject) => {
     .catch((err) => reject(err));
 });
 
-export default { getPins };
+const getBoardPinsbyBoardId = (boardpinBoardId) => new Promise((resolve, reject) => {
+  axios.get(`${baseUrl}/pins.json?orderBy="boardId"&equalTo="${boardpinBoardId}"`)
+    .then((response) => {
+      const boardPins = response.data;
+
+      resolve(boardPins);
+    })
+    .catch((err) => reject(err));
+});
+
+export default { getPins, getBoardPinsbyBoardId };
